refactoring new models
