--- app/Models/School/Branche.php
@@ -1,11 +1,0 @@
-<?php
-
-namespace App\Models\School;
-
-use Illuminate\Database\Eloquent\Factories\HasFactory;
-use Illuminate\Database\Eloquent\Model;
-
-class Branche extends Model
-{
-    use HasFactory;
-}
