+++ app/Models/School/Year.php
@@ -1,6 +1,6 @@
 <?php
 
-namespace App\Models;
+namespace App\Models\School;
 
 use Illuminate\Database\Eloquent\Model;
